Client / Server:  Simplified auth check / improved interceptor logic
